feat: ensure DeleteUnitOfMeasurementUseCase throws if no dependency is provided

+++ src/domain/usecases/unit-of-measurement/delete-unit-of-measurement.js
@@ -1,4 +1,4 @@
-const { MissingParamError } = require('../../../shared/errors')
+const { MissingParamError, ServerError } = require('../../../shared/errors')
 
 module.exports = class DeleteUnitOfMeasurementUseCase {
   constructor ({ unitOfMeasurementRepository } = {}) {
@@ -8,6 +8,14 @@ module.exports = class DeleteUnitOfMeasurementUseCase {
   async delete (id) {
     if (!id) {
       throw new MissingParamError('id')
+    }
+
+    if (!this.unitOfMeasurementRepository) {
+      throw new ServerError('No UnitOfMeasurementRepository provided')
+    }
+
+    if (!this.unitOfMeasurementRepository.delete) {
+      throw new ServerError('No valid UnitOfMeasurementRepository provided')
     }
 
     await this.unitOfMeasurementRepository.delete(id)
